move publish profile tests to their own file
