+++ extensions/sql-database-projects/src/test/baselines/publishProfileIntegratedSecurityBaseline.publish.xml
